+++ src/Service/S3/src/Input/GetObjectAclRequest.php
@@ -41,11 +41,20 @@ final class GetObjectAclRequest extends Input
     private $RequestPayer;
 
     /**
+     * The account id of the expected bucket owner. If the bucket is owned by a different account, the request will fail
+     * with an HTTP `403 (Access Denied)` error.
+     *
+     * @var string|null
+     */
+    private $ExpectedBucketOwner;
+
+    /**
      * @param array{
      *   Bucket?: string,
      *   Key?: string,
      *   VersionId?: string,
      *   RequestPayer?: RequestPayer::*,
+     *   ExpectedBucketOwner?: string,
      *   @region?: string,
      * } $input
      */
@@ -55,6 +64,7 @@ final class GetObjectAclRequest extends Input
         $this->Key = $input['Key'] ?? null;
         $this->VersionId = $input['VersionId'] ?? null;
         $this->RequestPayer = $input['RequestPayer'] ?? null;
+        $this->ExpectedBucketOwner = $input['ExpectedBucketOwner'] ?? null;
         parent::__construct($input);
     }
 
@@ -66,6 +76,11 @@ final class GetObjectAclRequest extends Input
     public function getBucket(): ?string
     {
         return $this->Bucket;
+    }
+
+    public function getExpectedBucketOwner(): ?string
+    {
+        return $this->ExpectedBucketOwner;
     }
 
     public function getKey(): ?string
@@ -99,6 +114,9 @@ final class GetObjectAclRequest extends Input
             }
             $headers['x-amz-request-payer'] = $this->RequestPayer;
         }
+        if (null !== $this->ExpectedBucketOwner) {
+            $headers['x-amz-expected-bucket-owner'] = $this->ExpectedBucketOwner;
+        }
 
         // Prepare query
         $query = [];
@@ -128,6 +146,13 @@ final class GetObjectAclRequest extends Input
     public function setBucket(?string $value): self
     {
         $this->Bucket = $value;
+
+        return $this;
+    }
+
+    public function setExpectedBucketOwner(?string $value): self
+    {
+        $this->ExpectedBucketOwner = $value;
 
         return $this;
     }