+++ src/Service/S3/src/Input/ListPartsRequest.php
@@ -57,6 +57,14 @@ final class ListPartsRequest extends Input
     private $RequestPayer;
 
     /**
+     * The account id of the expected bucket owner. If the bucket is owned by a different account, the request will fail
+     * with an HTTP `403 (Access Denied)` error.
+     *
+     * @var string|null
+     */
+    private $ExpectedBucketOwner;
+
+    /**
      * @param array{
      *   Bucket?: string,
      *   Key?: string,
@@ -64,6 +72,7 @@ final class ListPartsRequest extends Input
      *   PartNumberMarker?: int,
      *   UploadId?: string,
      *   RequestPayer?: RequestPayer::*,
+     *   ExpectedBucketOwner?: string,
      *   @region?: string,
      * } $input
      */
@@ -75,6 +84,7 @@ final class ListPartsRequest extends Input
         $this->PartNumberMarker = $input['PartNumberMarker'] ?? null;
         $this->UploadId = $input['UploadId'] ?? null;
         $this->RequestPayer = $input['RequestPayer'] ?? null;
+        $this->ExpectedBucketOwner = $input['ExpectedBucketOwner'] ?? null;
         parent::__construct($input);
     }
 
@@ -86,6 +96,11 @@ final class ListPartsRequest extends Input
     public function getBucket(): ?string
     {
         return $this->Bucket;
+    }
+
+    public function getExpectedBucketOwner(): ?string
+    {
+        return $this->ExpectedBucketOwner;
     }
 
     public function getKey(): ?string
@@ -129,6 +144,9 @@ final class ListPartsRequest extends Input
             }
             $headers['x-amz-request-payer'] = $this->RequestPayer;
         }
+        if (null !== $this->ExpectedBucketOwner) {
+            $headers['x-amz-expected-bucket-owner'] = $this->ExpectedBucketOwner;
+        }
 
         // Prepare query
         $query = [];
@@ -165,6 +183,13 @@ final class ListPartsRequest extends Input
     public function setBucket(?string $value): self
     {
         $this->Bucket = $value;
+
+        return $this;
+    }
+
+    public function setExpectedBucketOwner(?string $value): self
+    {
+        $this->ExpectedBucketOwner = $value;
 
         return $this;
     }